Fixes statsd mess and adds statsd k8s config.
This CL replaces the series of approved but abandoned CLs
that attempted to integration a statsd metric backend into
mixologist. It also adds statsd to the YAML for deployment.

It has been checked with "go test" in the top-level directory,
as well as the statsd/ dir. Additionally, a deployment to a
dev project was used to validate runtime.

Committing without review to patch broken repo (and based
on +2s for the abandoned series that gave way for this CL).

Change-Id: I3b6058aeb3187595c744ab643693a34d7614f125