Refactor logging integration to add adapter layer.
This CL refactors the initial glog integration to add a
notion of a logs adapter layer that translates logs to
an internal representation and handles dispatch to logging
backends. This should somewhat simplify logging backend integrations
moving forward.

Additional refactoring is likely necessary down the road to support
more advanced logging destination features (aggregation, batching).

The logs adapter comes with a simple default logger that writes to
stderr currently. This behavior is flag-controlled and can be
disabled.

Change-Id: I6ede2baaa5b2579434fff3df26bfd53a883831e7