Add initial stackdriver (cloud logging) integration.
This CL provides initial functionality for stackdriver logging.
To use, you must first establish a secret and mount it to the
mixologist that you have deployed. Only then can you enable
the stackdriver logger via the commandline.

Minimal test coverage is provided, but additional tests need to
be added later (coverage is ~60%).

Change-Id: I4e4d105e6ac35bdc47cc1fd0882609f9ff6c461d